+++ CHANGELOG.md
@@ -8,13 +8,13 @@ This project adheres to [Semantic Versioning](https://semver.org/).
 ### Added
 
 ### Changed
+- Improve the configuration check messages (#1061, #1062, #1063)
 
 ### Deprecated
 
 ### Removed
 
 ### Fixed
-- Improve the configuration check messages (#1061, #1062, #1063)
 
 ## 4.1.9