+++ CHANGELOG.md
@@ -6,7 +6,7 @@ This project adheres to [Semantic Versioning](https://semver.org/).
 ## x.y.z
 
 ### Added
-- Add an `AbstractConfigurationCheckViewHelper` (#1082)
+- Add an `AbstractConfigurationCheckViewHelper` (#1082, #1084)
 - Add a configuration-dependent validator (#1075, #1077, #1078, #1080)
 - Add a view helper for hiding template parts via configuration (#1070)